+++ packages/supacloud/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.8.17](https://github.com/zuohuadong/supacloud/compare/supacloud-v0.8.16...supacloud-v0.8.17) (2026-08-11)
+
+
+### Bug Fixes
+
+* **supacloud:** sync published CLI dependencies ([#830](https://github.com/zuohuadong/supacloud/issues/830)) ([29309d4](https://github.com/zuohuadong/supacloud/commit/29309d4cced5b5f997e85f2b03a523cf91b760f4))
+
+## [0.8.16](https://github.com/zuohuadong/supacloud/compare/supacloud-v0.8.15...supacloud-v0.8.16) (2026-08-11)
+
+
+### Bug Fixes
+
+* **supacloud:** sync published CLI dependencies ([#821](https://github.com/zuohuadong/supacloud/issues/821)) ([332bf1f](https://github.com/zuohuadong/supacloud/commit/332bf1f7b3c28edddcd177e06f8be3e2aaa4044d))
+
 ## [0.8.15](https://github.com/zuohuadong/supacloud/compare/supacloud-v0.8.14...supacloud-v0.8.15) (2026-08-11)